+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/useCases/AddClientToOrderUseCase.java
@@ -1,19 +1,18 @@
-package com.tech.challenge.tech_challenge.core.domain.services.extended;
+package com.tech.challenge.tech_challenge.core.domain.useCases;
 
 import com.tech.challenge.tech_challenge.adapters.driven.infra.repositories.OrderRepository;
+import com.tech.challenge.tech_challenge.adapters.driven.infra.repositories.PaymentRepository;
 import com.tech.challenge.tech_challenge.core.application.exceptions.ResourceNotFoundException;
 import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationException;
 import com.tech.challenge.tech_challenge.core.domain.entities.Client;
 import com.tech.challenge.tech_challenge.core.domain.entities.Order;
-import com.tech.challenge.tech_challenge.core.domain.services.OrderService;
-import com.tech.challenge.tech_challenge.core.domain.useCases.FindClientByIdUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
 
 @Service
-public class OrderClientService extends OrderService{
+public class AddClientToOrderUseCase {
 
     @Autowired
     private OrderRepository orderRepository;
@@ -21,8 +20,11 @@ public class OrderClientService extends OrderService{
     @Autowired
     private FindClientByIdUseCase findClientByIdUseCase;
 
-    public Order addClientToOrder(UUID orderId, UUID clientId) throws ValidationException, ResourceNotFoundException {
-        Order order = getById(orderId);
+    @Autowired
+    private FindOrderByIdUseCase findOrderByIdUseCase;
+
+    public Order execute(UUID orderId, UUID clientId) throws ValidationException, ResourceNotFoundException {
+        Order order = findOrderByIdUseCase.execute(orderId);
         Client client = findClientByIdUseCase.execute(clientId);
         client.validate();
