+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/useCases/FindOrderItemByIdUseCase.java
@@ -1,0 +1,17 @@
+package com.tech.challenge.tech_challenge.core.domain.useCases;
+
+import com.tech.challenge.tech_challenge.core.domain.entities.Order;
+import com.tech.challenge.tech_challenge.core.domain.entities.OrderItem;
+import org.springframework.stereotype.Service;
+
+import java.util.UUID;
+
+@Service
+public class FindOrderItemByIdUseCase {
+
+    public OrderItem execute(Order order, UUID orderItemId){
+        return order.getOrderItems().stream()
+                .filter(orderItem -> orderItem.getId().equals(orderItemId)).findFirst()
+                .orElseThrow();
+    }
+}
