+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/useCases/FindOrdersUseCase.java
@@ -1,0 +1,20 @@
+package com.tech.challenge.tech_challenge.core.domain.useCases;
+
+import com.tech.challenge.tech_challenge.adapters.driven.infra.repositories.OrderRepository;
+import com.tech.challenge.tech_challenge.core.domain.entities.Order;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class FindOrdersUseCase {
+
+    @Autowired
+    private OrderRepository orderRepository;
+
+    public List<Order> execute(){
+        return orderRepository.findAll();
+    }
+
+}
